Switch social auth provider name to github
While we didn't change auth provider in terms of external service
provider we did base the new provider class on Social Auth's
GithubOrganizationOAuth2 which uses the name `github-org`.  Changing
provider name means all the existing social auth database records are no
longer valid and logging in will generate a new user.  We tried to fix
this with a migration but found the dependency resolver was unhappy
about depending on both the jobserver and social_django apps because of
the Custom User Model, so this was the [much!] easier fix.