Remove comment around moving to ModelSerializer for Jobs
DRF's ModelSerializer class generates a Serializer class which is
further from our desired outcome than what we already have.  While it
would be nice to tie the API for Jobs, a table effectively controlled by
job-runner, to it's Model definition it turns out to be more work/lines
of code than defining exactly what we want with a plain Serializer.