Squash migrations
The move to a custom user model broke our migrations for emtpy databases
because the Admin's log table still had a constraint on auth_user.
Since we want our custom user to always exist when someone picks up the
project for the first time (or when we plug in the Admin) this squashes
all the migrations to one schema migration and one data migrations to
set up the Backends.