Ensure values used for filters exist in the db
Django does some automated conversation when we use the ORM so passing
in the string "123" to a filter, eg .filter(pk="123") will find a record
with the integer pk 123.  However if this value can't be converted to an
integer this filter will throw a ValueError.  Since our filter inputs
here are user generated this checks query is using valid values before
adding the filter to the query.