Refactor release files API
There is now just one file-serving api, but multiple index apis. This
allows each type of index to map names to specific files more flexibly.

We use this with the workspace index API to include the backend in the
name of the file, to make the same file name released in different
backends distinct.

Update the index API response to be an object that contains a list of
file metadata, which is more flexible and future proof.

Also, move the logic for listing the file work workspace from the model
into releases.py, as trying to contain all release logic there.