Refactor release upload API.
This is a refactoring of the Release and ReleaseFile models, and their
upload API.  Previously, a Release and its ReleaseFiles were created
atomically by single zip upload. This change separates creation of
a `Release` from uploading its `ReleaseFile`s, as well as tightens
various model contraints.

A user can now create a Release, and a different user can upload files
to that Release. This is to accomodate the intended workflow of
reviewing releases, and also the potential to upload some but not all
files requested.

In the process, some things have been simplified:

 - hashing and deduplication is now per file, rather than per zip, which
   was bugged.
 - removed `backend_user` and just use created_by, as we have a solution
   for that now on the TPP backend.
 - got rid of manifest uploading as it was not being used anywhere.
 - the API urls were simplified as a result and are more RESTish:
   - A single Workspace Releases API url, which on POST will create
     a release, and on GET will list workspace files.
   - A single RequestFile API url, which on PUT will upload and on GET
     will download.
   - A release index url, which GET will list files in release

Some things got more complex:

 - RequestFiles need to be uploading one by one, and validated that
   their filename, contents, and backend match the corresponding Release

Various other related changes:

 - Switch to UUID for Release.id
 - File listing API response changed from list of filenames:
    - top level dict with "files" key for security
    - files list of dicts for easier sorting
 - switched to sha256 to avoid md5 security silliness
 - fixed a ReleaseFile.get_absolute_path bug

What is not done in this PR, but will come later:

 - rewrite release management command to use new API
 - add permissions checks to uploading files
 - changing Release.status