Update release API endpoint
Note that although tests continue to pass (apart from the one we've
deleted), the behaviour's changed.  The middleware that catches the
PermissionDenied exception returns a 404, but the original code returned
a 403.

There's also a permission check in validate_upload_access (checking that
user is member of the given backend) that could/should be moved so that
it's defined on the action function.