Filter orgs based on existing memberships
When we show a list of orgs for a user to be added to we don't need to
show the orgs they're already members of.  This is already enforced by
OrgMemberships being unique on (org, user) but this avoids the error
condition before it happens and makes our lives easier.