Set workers in prod to 3, rather than the default of 1.
A single sync worker is usually not a good idea with gunicorn, as one
long request will block everything, and we're using whitenoise so there
are more requests for a pageload.

I temporarily added up to 4 workers in prod, but only 2 were every
really used, so 3 seemed a good start.