Check for applicant reaching the confirmation page
The form's flow currently only allows a user to edit their application
by clicking back or URL hacking.

This commit:

* Adds update buttons to the confirmation page so that someone can jump
  back to a particular form page.
* Tracks whether an application has reached the confirmation page.
* For applications that have reached the confirmation page, change the
  "Continue" button to return to the confirmation page.

Note that some amendments to the form (e.g. type of research) may result
in a user having to add further content to subsequent pages that they
have not reached before. That flow is a little disrupted here — a user
would still get passed back to the confirmation page.

For this initial version of the form, this is acceptable. We later
intend to validate the form entirely, so a user will see which other
parts of the form they will need to amend.