Fix bug with "internal error" notifications
Previously, if the job already existed it would notify based on the
_old_ status_mesage of the job, not the one to which we were about to
update.  This didn't get caught by the tests because if the job was a
new job then different update logic was used.

This PR amends the test to catch this condition and then slightly
refactors the update logic to fix the bug and hopefully make it clearer
(though its arguable whether the changes are an overall improvement).