Avoid setting blank shortNames for single file outputs responses
We pull a common prefix from the filenames served to the outputs viewer
app.  We use this prefix to create a `shortName` value on the files
array for display in the file list UI.  However with one file the
longest common prefix is the entire filename, leading to an empty
shortName, and showing an empty file list.