Add missing field
This was previously a field on Application.  It is special, in that it
should be shown on two different pages.  When we broken up the
Application model, I should have added it to both Page objects, but
didn't.

For now, a user will have to enter it once on each page,  which is not
ideal.