Add roles2 field to each model with a RolesField
Postgres array fields define the type of element they contain.  Django
exposes this as ArrayFields having a `base_type`, so we have two new
fields defined.  RolesArrayField for the array, and RoleField for the
strings in the array.

This splits up our logic nicely with RolesArrayField handling uniquing
of the input values and the RoleField handling the conversion between
role class and string.