Fix URL tests for Django 4.0
With the removal of update_wrapper in View.as_view() [1] CBVs no longer
have __qualname__ updated to match that of the view class.  Our
dotted_path() function makes use of __qualname__ when constructing the
path which we compare to the class symbol passed to the test.  So
instead it now sees the [correct] __qualname__ of the function built
inside View.as_view().  However we can still get the view class via the
view_class attribute, which is the expected method of getting it.

[1]: https://github.com/django/django/pull/14124