Update policy link to production site
In `snippets/terms.md`, the same page is linked, but on the production
site.

This change makes the links consistent.

The production site also has more content than the URL linked. For
example, the ONS-CIS information is "awaiting acknowledgement content"
in the page linked, but updated on the production site.