Explicitly set DRF authentication on relevant API endpoints
Instead of having SessionAuthentication as the default authentication
for all endpoints this stops all authentication by default, leaving the
choice of authentication method up to each endpoint.  We have multiple
authentication (and authorization) schemes in play and this attempts to
make it clearer what is being used by a given endpoint.