Ignore missing coverage of GraphQL pagination handler
We use GitHub's GraphQL endpoint for two queries, both related to the
repos of the `researcher` team in the `opensafely` organisation.  As
such the _get_query_page method flattens the initial part of the
response hierarchy for us so the calling methods don't have to all drill
into

    results["data"]["organization"]["team"]["repositories"]

This is also useful because we only paginate on the repositories part of
those queries and don't want to have to dynamically look for and then
consume paginated resources.

However, this makes verifying our pagination code tricky without
generating 100 repos in our test org.

So I've taken the pragmatic step here and ignored coverage for it.