Update private repos report page to correctly filter out repos
This is a full rework of the view essentially.  The major change is how
we treat the data.  Repos are now the first class object throughout the
code, they get enhanced with workspaces and related derived data.  All
filtering is done on the enhanced repo structure.

I've dropped the project column since it can be reached via the
Workspace.