check for existance of files using uploaded_at
When a ReleaseFile has been created but no file uploaded we don't have a
path for it, so ReleaseFile.absolute_path() will return a valid path to
the releases directory, which will exist.  However a file being on disk
isn't the most correct check here since a file could have also been
deleted from disk, so lets check to see if we think we've uploaded it
previously.