feat: serve helpful content to SPA for missing files
ReleaseFiles might not be able to be served: if they have been redacted,
if they have not yet been uploaded, and if the file is simply not there.

Currently, all the above cases, the API returns a 404, for which the SPA
displays a generic page and pop up error. This doesn't give the user
much clue as to why the file cannot be served.

This change serves place holder content, specific to each of the above
cases, which informs the user why they cannot view the file contents.

 - it returns these as 200s, rather than the more correct 404, as the
   SPA will error on anything else.
 - it returns a plain text message. This might be prettier as a html
   template with links, perhaps.

Problem: it requires the SPA to use the content-type from the response
to decide what viewer to user.