fix: view not-yet-uploaded files for a release
Currently, there is no way to via a release with files from
release-hatch. All the release metadata is stored in jobserver, but the
contents are stored in release-hatch, so the SPA really needs data from
both sources.

This change is a quick hack to redirect to release-hatch for the file
contents if the file has not been uploaded. This will only work on level
4, but provides a basic view of the pending release for an OutputChecker
to look at.

It works by returning an empty 200 with some headers, including auth,
which the output-viewer manually fetches.

It would be better to add support for this mode of viewing to the
output-viewer itself, and launched in a specific manner different to the
current method. But this will work for the purposes of a demo.