Handle uploaded file's hash not matching the requested ReleaseFile.filehash
We want to feed this back to the user and raising a ValidationError is
the correct way to do this in a DRF world.  We don't want to couple the
releases module to DRF though so as with ReleaseFileAlreadyExists we're
using a custom exception and capturing it outside that module.