feat: add new honeycomb direct links
This change adds honeycomb direct links to the trace for a job.

To do this, job-runner needs to send us the trace context, which we're
not doing yet, but will once this change is deployed.

I've removed all the current honeycomb links, as they are based on the
old tracing data. I've also refactored the code for rendering those
links, to make it easier to add new onces without changing the template.

We will probabaly add more template links again, once we have a better
idea of what would be useful.

This change also explicit sets up tracing in tests, same as how we've
done it in job-runner. We don't actually use this for any tests yet, as
we do not emit any manual traces to test at the minute. But it was
needed in order to be able to have the Job factory generate a suitable
mock trace_context.

Fixes #2159