Django management command to delete empty projects
We've been asked to [delete a
project](https://github.com/opensafely-core/job-server/issues/1991) that no
longer has any workspaces associated with it, as part of a data tidying
exercise for older work in job server.

There is currently no way to do this in the UI (e.g via the staff area), so
this management command adds a quick way for admins to do this without manually
editing the database.