Look up file counts for copilot dashboard in a second query
When annotating the number of files for a project the query projects
query jumps from ~120ms to ~20s locally.  I can't see anything obviously
wrong from the query planner and it seems to specifically be the files
annotation causing this slow down.  So instead this does a separate
query, grouping in memory by project and coverting to a dict of
project_pk: number_of_files, which is merged with the project data.