Ignore jobs associated with unknown JobRequests
We're planning to update the job-runner sync protocol such that it syncs
the state of all active jobs, where active now means those jobs which
job-server thinks are active _and_ those which job-runner thinks are
active. This should make the sync process even more robust against
"non-standard" things happening (i.e. us fiddling with the state of
jobs).

This changes means that it's no longer a protocol violation for
job-runner to tell job-server about a job which job-server didn't
explicity ask about. If this happens (which I don't expect it to really)
we should log it but not treat it as an error.