Notify for newly completed jobs with no previous state
Unless we do this then jobs which transition immediately into an error
state (such as those produced by a JobRequestError) will never notify
and so we won't get Sentry alerts about them.

The previous worry, that we might end up notifying about jobs we had
already notified about seems unfounded, and in any case not a disaster
if it were to happen.