Fix gunicorn bind config and simplify smoke-test
The main problem with the prod image was gunicorn binding to 127.0.01
rather than 0.0.0.0, which this change fixes.

However, this was tricky to diagnose, as it was hard to run the prod
image locally. To make this easier, this change refactors the
docker-compose stack and justfile to support running prod image in the
same way as dev.

This means that all docker justfile targets (except test) now take an
initial env arg, which defaults to `dev`. This means that we can then do
`just serve prod` and it will work.

For the docker-compose prod service to be able run locally, it needed to
use the correct DB config, and also depend on the db server being
started. So I added a new `base` service that sets the DATABASE_URL
config for *all* compose services, and allows us to have a small `prod`
service defined that depends on the `db` service, similar to how we did
dev-base (which now extends `base`)

Now that it was working the smoke-test was greatly simplified, so
I added a smoke-test just target that just runs curl -I, and takes
a host parameter. On linux, the default value of http://localhost:8000,
but on OSX you might need to do:

`just docker-smoke-test http://$DOCKER_VM_IP:8000`

We can add to what that target does in future.

And finally, I combined the test and smoke-test docker workflows
together, as the changes meant the smoke test was so simple and fast
that it wasn't worth parallelising.

I have some plans for refactoring that further, but I wanted to limit
the size of this change, which is already a bit large.