chore: Sort docker static assets processing
This refactors collectstatic tooling and adds
scripts/collect-me-maybe.sh, which is runs collectstatic if any assets
*or* static files have changed. Previous set up only detected asset
changes, not static files.

Putting it in a script allows us to use it in both host-based and
docker-based tooling.

Relatedly, this changes makes STATIC_ROOT configurable via env var, and
uses this to configure the docker env to use a different directory for
staticfiles. This separates the host and docker envs, with each owning
separate directories for their assets build and static files.