Remove published outputs for a workspace badge
We've moved away from exposing data in job-server to readmes and are
instead focussing on driving viewers of readmes to the appropriate
project and exposing all relevant data they might need there, so there's
no intention to use badges any more.

This also removes our only use of the deprecated imghdr from the stdlib,
removing a blocker to deploying to 3.11.