Maintain redirect suffixes
We match redirects using, effectively, a startswith() call.  However on
a match we were redirecting to the linked objects URL.  This works when
someone is trying to view that object (eg a project), but if they're
trying to look at a child object (eg a job) they were still ending up on
the object page.