handle invalid configs
We spent nearly 18mo accepting jobs with config that didn't go through
validation until it was parsed by job-runner.  Errors were caught there
and fed back to the user via our sync mechanism.  Now that we expose the
full run command of a job to the user we're parsing configs on page
load.  For those older jobs this can sometimes cause an error.  I've
found ~16 workspaces where this is an issue so I think we can just not
display the run command for those.