display a Project's application_url when one is set
A Project has either:

* an Application via FK
* an application_url
* neither

so we need to display relevant buttons or links depending on that state.
There's no process here for a Project missing both and someone wanting
to add an application_url because we don't currently want to support
that since that use case is intended to be covered by the user
management form.