Wrap getting the content of AnalysisRequest.report in a property
An AnalysisRequest's report can be missing in a couple of different ways
but when we consume that reports content we don't want to worry about
those reasons at call time.  Either the report has content or it
doesn't so we wrap up the content lookup in a property that always
returns a string of some sort.