rebuild AnalysisRequest model with ULID PK
Timeflake used the UUID column type under the hood in postgres.  The
TimeflakePrimaryKeyBinary field parsed any query values such that ORM
constructions like `filter(pk="blah")` and `.get(pk="blah")`, that is
strings with invalid timeflake values in them, raised a ValidationError
because they could not be parsed into a timeflake object.  This is both
consistent with other Django field types such as UUID (which tries to
instantiate a stdlib uuid.UUID object with the given value), but
unexpected when one is used to working with AutoField.  It also makes
common usage patterns of the pk much harder to work with.  When querying
using the `filter(pk="blah").first()` pattern, one cannot rely on
getting an object or None, as the value passed to pk may raise a
ValidationError, requiring further handling of the query.