Replace the navs list-o-dicts with objects to support predicates
We have a number of use cases for optional menu items.  To test those
out this adds a way to add items to the top level navs with an optional
predicate which is checked when outputing the structure for the
template.

The staff area nav has been included in this change for consistency,
although we currently don't expect to use predicates for that.