fix disabling of the internal repo sign off button
The {% if disabled %} check in the template was incorrect because a dict
with any keys is always truthy.  So when the button should have been
enabled it was disabled with an empty tooltip because none of the keys
were True.

I've switched over to an object with a modified __bool__ implementation
that aggregates the boolean values of each instance variable so we can
make that check in the templates.

This also fixes a small bug in the template logic which caused both the
disabled and enabled buttons to appear.