chore: Fix using requirements.prod.txt as constraints
Newer pip doesn't like `[extras]` in constraints files. We are using
our `requirements.prod.txt` as a constraints file for our dev deps, to
ensure any shared transient deps are in sync.

Dependabot is using a newer pip that chokes on this, and is thus
currently failing:

https://github.com/opensafely-core/job-server/security/dependabot/29/update-logs/319978516

There's no real need for extras to be in the package names in a .txt
lockfile, as the pip-compile step will fully resolve any deps from
extras in the .in file ahead of time anyway. So we use the
`--skip-extras` to remove them from the .txt file.

However, this exposed us to a bug when using required hashes,
contstraints, and strip extras together:

https://github.com/pypa/pip/issues/9644.

Basically, pip does not parse the hashes properly from the extra-less
name in the constraints file, leading to it to fail installing.

There's a simple work around, just use `--no-deps` when installing. As
pip-compile has fully resolved all the dependencies ahead of time, then
we don't need pip to do it again at install time.

In turns out, this is actually recommend practice when using hashes, as
it avoids accidentaly installing things that are not specified:

https://pip.pypa.io/en/stable/topics/secure-installs/#do-not-use-setuptools-directly

Whether this actually fixes dependabot or not remains to be seen.