Move analysis template context to JSONField
We are expecting to have more than one type of analysis in the
medium-to-long term.  This is a step towards that.  We don't need the
specific fields submitted by an analysis as "first class" database
fields, and keeping them in JSON allows us to avoid having to migrate
their schema over time.