fix: Merge code used to actually render the current templates
This should work with the main branch of interactive-templates, when the
current PR lands with it.

This method assumes the templates for a specific interactive analysis
are in a directory call `templates/NAME`, where name is hardcoded to
'v2' for now. The files in that directory are recursively
copied/rendered as needed into the target repo. This simplifies the
`create_commit()` function quite a bit.

Expanded the default template_repo fixture to include some more files,
to properly exercise the templating.