Add buttons linking to the interactive form
We expect the happy path for Interactive users to be login -> redirect
to the interactive form.  However it's possible they will end up on the
appropriate Project/Workspace page, these buttons will give them a route
to get to the form.

It will also benefit us while testing and debugging where we, as staff,
won't be following the same path as the typical user, so having a clear
route to get to the interactive form will be useful.