feat: add interactive_templates as a package
Now we just import the schema from the `interactive_templates` package,
and call its render method.  The removes all references and use of
a template repo, and tests that convered the actualy rendering logic.

As the interactive_templates packaging branch cannot land until this
change is deployed, we pin to specific branch commit of it in this
change. After this is deployed, we'll switch the dependency to point to
a version tag on main.