Remove unrelated fields from user page for email-only users
Our email-only users don't get access to various related objects or
fields, so the staff area's UserDetail page was a little confusing for
those users.

This splits the page into two views, one for users who sign in with
OAuth and one for users who sign in with email.  Rather than splitting
up the URLs and causing confusion with which type to use there is a base
UserDetail view which handles deferring to the correct view for the type
of user.