Only notify output checkers when the JobRequest has succeeded
We only want to create a ticket for output checkers when the JobRequest
enters the successful state, rather than any single job entering a
completed state.

We have to do some leg work to look up the status both before and after
the Job updates because we're dealing with JobRequests which come from a
group_by generator, rather than a QuerySet.