Fix queries on roles fields
We typically only access roles fields via the helper functions provided
in the authorizations package so querying them directly hasn't come up
much before.  However, when one is trying to filter for models with a
particular role using __contains and passing in the given role we still
want that to work.  The RolesField implementation expects an iterable
when preparing values so we cast that when we only get one now.