Link Project.number uniqueness failures to their form field
When a constraint fails a ModelForm puts the error into the
non_field_errors, which we display at the top of our form.  The
IntegrityError is also raised somewhere in the ModelForm's clean stack
(originally coming from the link model's clean stack).

Since this failure is specifically for the number field that feels quite
confusing in the UI.

This change moves ProjectCreateForm to a plain Form so we can handle the
validation of the model where we create it (sticking to the principle of
locality of behaviour), and thus attach the error to the correct field
on our form instance allowing our error handling to display it next to
the number field.