chore: add basic login token functionality to User model
`User.generate_login_token` will generate a single (token, expiry) pair
and save it. The token itself is hashed.

`User.validate_login_token` will validate that the provided token is
valid, and then delete it if so.

Whitespace is striped from the token, to allow for formatting for users.

This could live somewhere else except the User model, but it did seem
natural fit, and similar to the current PAT tokens we support.