Replace OAuth authentication pipeline with a single function
PSA has been very useful to us, but its near infinite level of
configurability does come at a cost when trying to reason about what
state a user will be in.  One of several examples of this was us having
spend time working out how to stop it from overwriting User's name
fields.

We also have 3 methods of login to the site now.  Each of these methods
uses a different set of fields to look up User[SocialAuth] records being
explicit about which model and fields are used for that will aid in
maintainability.

This takes the relevant parts of our old pipeline and converts them to a
single function which roughly follows this flow:

1. look up UserSocialAuth record with the UID from GitHub
2. if we have one, the user is logged in
3. if we don't, create the UserSocialAuth and User instances
4. do any post-creation steps

This change has allowed us to remove our functions to set fullname and
notifications_email because they're now inherently handled as part of
creating a new user.