Switch to two templates for the homepage
This splits the homepage into two templates, based on a viewers current
authentication state, making it easier to jump to those templates when
coming from the view.

This also pulls the job request list structure fully into the partial
for rendering it since that's the only common part of the two pages.