Capture a name for users who don't have one set on GitHub
Python Social Auth's pipeline feature has the concept of pausing and
returning to a run through it [1].  This lets us eject the user to eg
capture some more information in a traditional view.  We're doing
exactly that here by showing them a form to get their name if they
haven't set one in GitHub.

[1]: https://python-social-auth.readthedocs.io/en/latest/pipeline.html#partial-pipeline