Tidy wording of the Interactive welcome email
This is the email that users receive when their account is added to Job Server.
No more set up is required once it's created, so they should just be able to
log in and start an analysis.

I also noticed the one-time link to login didn't have a closing tag in the
login.html template, so have added that.