Shift Django/interactive-templates boundary
We are starting to capture values in the react form that are not
necessary for rendering an analyses' code from the templates.  In light
of that this switches us from doing:

1. parse data into Analysis structure
2. re-shape data for AnalysisRequestForm
3. validate data with AnalysisRequestForm instance

to using the form first:

1. parse data into AnalysisRequestForm
2. validate data with AnalysisRequestForm instance
3. build an Analysis structure from the validated data

This helps us maintain a boundary where the Analysis structure is only
for data used to render the analysis template, and job-server breaks out
the relevant incoming data to build that structure.

Note: purpose needs to be removed from Analysis but will come in a
future commit.