Forceably reinstall interactive templates package
Locally, changes to the interactive_templates package are not picked up
automatically because we are usng a zipfile rather than providing an explicit
version. This means we need to uninstall and reinstall the package to pick up
any changes.

This is not a problem in CI because it is building from a fresh
environment each time.

Note: this change should be removed when the package has proper versions that
pip can handle appropriately. This is not intended to be a long term fix.