Calculate the start of the last full week of data
This calculates the Monday of the week with the most recent complete extract of
data from TPP. This is passed into the Analysis and used to calculate the
number of people with matching events in that week. This number is then shown
on the report (with low number redaction).

Fixes interactive-templates#17