Add decision fields to publish request models
We're transitioning from storing approvals, with rejection being
implicitly tied to pending, to explicity approved/pending/rejected
states.

We're storing these on both models at this point because we want a
single place to track the publish request for a report, but that
inherently requires tracking the same for the ReleaseFile it's built
from.  In OSI this is hidden from the user, but for the rest of the
platform it can exist without a related Report.

A future change might be to keep the decisions for publishing on
ReleaseFilePublishRequest if we felt ReportPublishRequest didn't gain us
anything but we're not sure if we can do that currently.