Specify report via command line argument
The report to be released by `osi_release` is now specified via a command line
argument rather than being inferred. This is usually going to be the report
outputted by the `osi_run` command, but it defaults to a fake report to avoid
the need to run an analysis.

The combined `osi_run_and_release` command has been updated so that the report
generated by `osi_run` is explicitly passed into the `osi_release` command.