Rework enabled/disabled state of publish request button for reports
Now that we can have both multiple publish requests for a given report
but only one pending request at a time we need to reflect that in the
button states we present to the user.

To simplify the complex logic necessary for these states, and to provide
explicit reasons for disabling the button, ReportPublishRequest exposes
its various states as properties.