Switch existing SERIAL columns to IDENTITY ones
Postgres introduced sql-standard IDENTITY syntax for columns in v10, and
Django has switched to using those for when creating implicit PK fields
as of 4.1.

The main draw of IDENTITY columns is not having a separate sequence for
the column.

The main draw of doing this for us is to have all of our [implicit] PKs
created in the same way, and to avoid drift between prod and non-prod
databases.  It's very possible this will never bite us but this is a
defensive change just in case.

Adam's article explains more and is the source of this management
command:

https://adamj.eu/tech/2022/10/21/migrate-postgresql-ids-serial-identity-django-4.1/