Drop use of whitenoises manifest config
After django-vite switched to looking up the URL for assets via the
staticfiles subsystem [1] all our tests which checked the
rendered_content (ie the rendered template) were failing with files not
being found from the manifest.  Local development with runserver was
unaffected.

We decided to keep the compression storage backend and ditch the
manifest one.

1: https://github.com/MrBin99/django-vite/commit/0e8c9ff80513c5f2326903a6c9ae4eb33e3dbcd8