Use release_file_view permission to gate access to viewing an analysis
We want staff members to be able to see unpublished reports.  They
typically have the ProjectCollaborator role applied globally.

A report is the main object being viewed with the AnalysisRequestDetail
view.

So it makes sense to switch the permissions check to use the
release_file_view permission and add that to the InteractiveReporter
role.

This will allow:
* users with the InteractiveReporter role to continue to view the
  outcome of their analysis requests
* staff members to view analysis requests on projects they are not
  members of