Link to the report page from the publication review ticket
The process for approving/denying a publish request is very simple,
click either approve or reject.  Rather than direct the user to the team
manual and then have them go look for the report either on the ticket or
via the search/list pages we can send them directly to it.  The
instructions on what happens now live right by the controls themselves.