Test all icon tags in our templates are valid tags
We don't test every template fully renders because of the time penalty
rendering templates imposes on the test suite.  The downside of this
approach is sometimes there are template-only things which need testing.
Typically one of those failures results in a test case where the
rendered template is checked.  However, for icons this would place an
overhead on frontend/UI development which we don't think is useful.

This test finds every icon tag definition across all our templates, and
checks those names are defined in the components configuration file.
This is intended to catch spelling mistakes or icon renames.