Upload the HTML coverage report when coverage fails
Sometimes we get coverage failures in CI, and it can be useful to have a
quick pointer as to why without running the test suite locally.  This
uploads the HTML output from coverage.py to GitHub if any has been
generated.  It only runs if the previous step has failed, which <100%
coverage will do.  It's ignoring a missing htmlcov directory because we
know that will happen when coverage is 100% or doesn't run because a
test failed.