Update application.md
As the OpenSAFELY service has now changed names, the wording on the start page of the application form needs to change from OpenSAFELY-EMIS/-TPP COVID-19 research platform to NHS England OpenSAFELY COVID-19 service. 

A hyperlink to this new direction (https://digital.nhs.uk/about-nhs-digital/corporate-information-and-documents/directions-and-data-provision-notices/data-provision-notices-dpns/opensafely-covid-19-service-data-provision-notice) will also need to be included to the wording at the top of the form where it first says 'NHS England OpenSAFELY COVID-19 Service'.