Add redirects for all projects to remove org from their URLs
This migration is a little non-standard.  It needs to build a URL
without using reverse (for reasons explained in the docstring), and
redirects need to be created by someone.  Since I, George, am the author
of these redirects I've set my user object so our audit trail is
correct.  However, my user does not exist in a fresh database such as in
CI, hence the oldest user lookup.